win: allow fatal errors to trigger WER (#3838)
In order to properly generate minidumps during actual events, it is
necessary to disable the call to `SetErrorMode()` during libuv
initialization and to remove the job flag
`JOB_OBJECT_LIMIT_DIE_ON_UNHANDLED_EXCEPTION`, both of which interfered
in the normal handling of exceptions. Finally, it was determined
prudent to include `CREATE_DEFAULT_ERROR_MODE` when spawning new
processes, to ensure that whatever error handling mode was set in the
current process is not inherited by subprocesses.

This aligns better with coredump handling on unix, where it is the
responsibility of the embedding application or user to specify the
desired behavior here.