unix: mark visibility hidden for function
A very slight performance optimization: the `-fvisibility=hidden` only
applies to function definitions, while explicitly specified visibility
attributes and pragmas applies also to declarations, avoiding the need
for PLT entries for these cases. This optimization is already the
default when compiling for Windows (which requires marking explicitly a
definition with dllimport to turn off this optimization).