misc: add clang thread safety analysis support
Implements declarations for being able to use clang's thread safety
analysis in client libraries that utilize uv's mutexes. Also enable them
for building libuv itself. This is a static version of the Thread
Sanitizer tests that are already in use on Unix. We will need to add a
clang CI agent for Windows to enable the testing there.